% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safesky.R
\name{live_get}
\alias{live_get}
\title{Reads data from safesky, updates the cache and sends live telegram alerts}
\usage{
live_get(pings_source = "puretrack")
}
\description{
Reads data from safesky, updates the cache and sends live telegram alerts
}
