% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{message_everyone}
\alias{message_everyone}
\title{Send a broadcast message to all telegram groups}
\usage{
message_everyone(message, override_daylight)
}
\arguments{
\item{message}{}
}
\value{

}
\description{
Send a broadcast message to all telegram groups
}
\examples{
#message_everyone("The alerts app has been updated")
}
