% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safesky.R
\name{read_safesky_live}
\alias{read_safesky_live}
\title{Reads data from safesky, updates the cache and sends live telegram alerts}
\usage{
read_safesky_live()
}
\description{
Reads data from safesky, updates the cache and sends live telegram alerts
}
