% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_site_pings}
\alias{summarise_site_pings}
\title{<b>{nearest_site_name}</b>\n{summary_text}\n<a href="https://live.glidernet.org/#c={lat},{long}&z=13&m=4&s=1&w=0&n=0">glidernet map</a>'}
\usage{
summarise_site_pings(pings)
}
\description{
<b>{nearest_site_name}</b>\n{summary_text}\n<a href="https://live.glidernet.org/#c={lat},{long}&z=13&m=4&s=1&w=0&n=0">glidernet map</a>'
}
